requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0

PyAudio>=0.2.11
sounddevice>=0.4.4
soundfile>=0.10.3
numpy>=1.19.5
soundcard>=0.4.2

faster-whisper>=0.9.0

tkinter
Pillow>=9.0.0
pywin32>=228
python-ctypes>=0.0.2

pygetwindow>=0.0.9